requirements.txt
beautifulsoup4
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
Pillow
piexif
geopy
janome
pykakasi
